+++ test/ggity_theme_test.exs
@@ -1,6 +1,7 @@
 defmodule GGityThemeTest do
   use ExUnit.Case
 
+  import GGity.Element.{Line, Rect, Text}
   alias GGity.{Plot, Theme}
 
   setup do
@@ -26,6 +27,32 @@ defmodule GGityThemeTest do
       assert String.contains?(stylesheet, ".gg-text {font-family: Helvetica")
       refute String.contains?(stylesheet, ".gg-axis-line")
       refute String.contains?(stylesheet, "height")
+    end
+
+    test "removes invalid data for line elements", %{plot: plot} do
+      plot = Plot.theme(plot, axis_line: element_line(size: "'22'<script>hackety-hack</script>"))
+      stylesheet = Theme.to_stylesheet(plot.theme, "gg-1") |> IO.chardata_to_string()
+      refute String.contains?(stylesheet, "22")
+      refute String.contains?(stylesheet, "script")
+      refute String.contains?(stylesheet, "hack")
+    end
+
+    test "ignores invalid data for rect elements", %{plot: plot} do
+      plot =
+        Plot.theme(plot, panel_background: element_rect(fill: "'22'<script>hackety-hack</script>"))
+
+      stylesheet = Theme.to_stylesheet(plot.theme, "gg-1") |> IO.chardata_to_string()
+      refute String.contains?(stylesheet, "22")
+      refute String.contains?(stylesheet, "script")
+      refute String.contains?(stylesheet, "hack")
+    end
+
+    test "ignores invalid data for text elements", %{plot: plot} do
+      plot = Plot.theme(plot, text: element_text(face: "'22'<script>hackety-hack</script>"))
+      stylesheet = Theme.to_stylesheet(plot.theme, "gg-1") |> IO.chardata_to_string()
+      refute String.contains?(stylesheet, "22")
+      refute String.contains?(stylesheet, "script")
+      refute String.contains?(stylesheet, "hack")
     end
   end
 end